+++ src/provider/SemanticTokensProvider/tools/funcHighlight.ts
@@ -5,7 +5,7 @@ import { getGuiFunc } from '../../../tools/Command/GuiTools';
 import { getHotkeyWrap } from '../../../tools/Command/HotkeyTools';
 import { getMenuFunc } from '../../../tools/Command/MenuTools';
 import { getSetTimerWrap } from '../../../tools/Command/SetTimerTools';
-import { getSortFunc } from '../../../tools/Command/sotrTools';
+import { getSortFunc } from '../../../tools/Command/sortTools';
 import type { TScanData } from '../../../tools/DeepAnalysis/FnVar/def/spiltCommandAll';
 import type { TSemanticTokensLeaf } from '../tools';
 import { regFnHighlight } from './regFnHighlight';
